v0.2.9 (4)
- Code refactoring
- Added errors logging
- Stats view filters optimizing
