+++ financecontrol/Extensions/DateExtensions.swift
@@ -8,10 +8,6 @@
 import Foundation
 
 extension Date {
-    var lastWeek: Date {
-        return Calendar.current.date(byAdding: .day, value: -7, to: self)!
-    }
-    
     var previousDay: Date {
         return Calendar.current.date(byAdding: .day, value: -1, to: self)!
     }
